refactor(core): remove custom .GEL parent class scope solution, new global normalize implementation via opt-in normalize prop, remove noReset and noScope props, update test examples
