+++ components/core/package.json
@@ -61,7 +61,6 @@
 	},
 	"dependencies": {
 		"@babel/runtime": "^7.9.6",
-		"@emotion/cache": "^10.0.29",
 		"@emotion/core": "^10.0.28",
 		"@emotion/weak-memoize": "^0.2.5",
 		"classnames": "^2.2.6",
