+++ components/core/src/GEL.js
@@ -7,20 +7,20 @@ import { BrandContext } from './Brand';
 import { useFocus } from './useFocus';
 import { Core } from './Core';
 
-export const GEL = ({ brand, noReset, noScope, children, ...props }) => {
+export const GEL = ({ brand, normalize, children, ...props }) => {
 	useFocus();
 
 	return (
 		<BrandContext.Provider value={brand} {...props}>
-			<Core noReset={noReset} noScope={noScope}>
-				{children}
-			</Core>
+			<Core normalize={normalize}>{children}</Core>
 		</BrandContext.Provider>
 	);
 };
 
 GEL.propTypes = {
-	noReset: PropTypes.bool,
-	noScope: PropTypes.bool,
+	normalize: PropTypes.bool,
 	brand: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),
+};
+GEL.defaultProps = {
+	normalize: false,
 };
