+++ package.json
@@ -90,7 +90,6 @@
 		"@babel/preset-env": "^7.9.6",
 		"@babel/preset-react": "^7.9.4",
 		"@changesets/cli": "^2.6.5",
-		"@emotion/cache": "^10.0.29",
 		"@emotion/core": "^10.0.28",
 		"@emotion/weak-memoize": "^0.2.5",
 		"@manypkg/cli": "^0.13.0",
